refactor(gateway): split planner and session/prompt along natural seams
planner/index.ts was 653 lines mixing three responsibilities: the public
PlanRequest schemas + Service/layer, the per-(peer,skill) tool factory, and
a bundle of pure helpers (name normalization, verified-label logic,
<remote_content> envelope, json-schema→zod). Split into:

  planner/index.ts      297  schemas + Service + layer
  planner/skill-tool.ts 216  buildSkillTool + padToolDescription
  planner/util.ts       184  pure helpers, re-exported from index

session/prompt.ts shed its bottom 70 lines of buildSystemPrompt,
mapFinishReason, evtUsage, and wrapTool into session/prompt-util.ts so
the layer factory reads top-to-bottom without scrolling past helpers.

Typecheck clean, 216/216 tests pass. No behavior changes — tests
consuming `../planner` exports still work via re-exports in index.ts.
